+++ app/jobs/drupal_import_meeting_job.rb
@@ -1,0 +1,9 @@
+# frozen_string_literal: true
+
+class DrupalImportMeetingJob < ApplicationJob
+  queue_as :exports
+
+  def perform(*args)
+    Decidim::DrupalMeetingsImporterService.run(*args)
+  end
+end
